issue/010: Add fixes

+++ src/main/java/org/spring/services/items/StrategyServiceItemDelete.java
@@ -1,11 +1,15 @@
 package org.spring.services.items;
 
+import org.spring.models.SaleItem;
 import org.spring.repository.RepositoryItem;
+import org.spring.repository.RepositorySaleItem;
 import org.spring.responses.ResponseBuilderItem;
 import org.spring.services.StrategyService;
 import org.spring.responses.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 /**
  * Service.
@@ -16,16 +20,23 @@ public class StrategyServiceItemDelete implements StrategyService {
     @Autowired
     private RepositoryItem repositoryItem;
 
+    @Autowired
+    private RepositorySaleItem repositorySaleItem;
+
     /**
      * {@inheritDoc}
      */
     @Override
     public Response getResponse() {
         if (repositoryItem.findById(HelperItem.getId()).orElse(null) != null) {
-            repositoryItem.deleteById(HelperItem.getId());
-            return ResponseBuilderItem.getResponseOkForDelete();
-        } else {
-            return ResponseBuilderItem.getResponseNotFound();
+            List<SaleItem> listSales = this.repositorySaleItem.findItemRelated(HelperItem.getId());
+            if (listSales.isEmpty()) {
+                repositoryItem.deleteById(HelperItem.getId());
+                return ResponseBuilderItem.getResponseOkForDelete();
+            }
+            return ResponseBuilderItem.getResponseConflictEntityHasRelations();
         }
+        return ResponseBuilderItem.getResponseNotFound();
+
     }
 }
