+++ src/main/java/org/spring/services/users/StrategyServiceUserPut.java
@@ -1,12 +1,16 @@
 package org.spring.services.users;
 
+import org.spring.models.Sale;
 import org.spring.models.User;
+import org.spring.repository.RepositorySale;
 import org.spring.repository.RepositoryUser;
 import org.spring.responses.ResponseBuilderUser;
 import org.spring.services.StrategyService;
 import org.spring.responses.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 /**
  * Service.
@@ -17,21 +21,32 @@ public class StrategyServiceUserPut implements StrategyService {
     @Autowired
     private RepositoryUser repositoryUser;
 
+    @Autowired
+    private RepositorySale repositorySale;
+
     /**
      * {@inheritDoc}
      */
     @Override
     public Response getResponse() {
         User user = repositoryUser.findById(HelperUser.getId()).orElse(null);
+        HelperUser.getEmptyList().add(HelperUser.getEntity());
+        // If user not exist.
         if (user != null) {
+            // If user have not equals ids.
             if (!HelperUser.getId().equals(HelperUser.getEntity().getId())) {
-                repositoryUser.deleteById(HelperUser.getId());
+                List<Sale> listSales = this.repositorySale.findSalesRelatedWithUser(HelperUser.getId());
+                // If user not have a relations
+                if (listSales.isEmpty()) {
+                    repositoryUser.deleteById(HelperUser.getId());
+                    repositoryUser.save(HelperUser.getEntity());
+                    return ResponseBuilderUser.getResponseOkForPut();
+                }
+                return ResponseBuilderUser.getResponseConflictEntityHasRelations();
             }
             repositoryUser.save(HelperUser.getEntity());
-            HelperUser.getEmptyList().add(HelperUser.getEntity());
             return ResponseBuilderUser.getResponseOkForPut();
-        } else {
-            return ResponseBuilderUser.getResponseNotFound();
         }
+        return ResponseBuilderUser.getResponseNotFound();
     }
 }
